// src/api/openAPIDefinition.schemas.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
export interface BookingRequestDTO {
  flightId: number;
  userId: number;
  passengers: PassengerInfoDTO[];
}

export interface PassengerInfoDTO {
  firstName: string;
  lastName: string;
  email: string;
}

export interface ApiResponseBookingResponseDTO {
  success?: boolean;
  code?: number;
  message?: string;
  data?: BookingResponseDTO;
}

export interface BookingResponseDTO {
  id?: number;
  reference?: string;
  status?: string;
  userId?: number;
  flightId?: number;
  bookingTime?: string;
  totalPrice?: number;
  passengers?: PassengerInfoDTO[];
}

export interface RegisterRequestDTO {
  email: string;
  /**
   * @minLength 8
   * @maxLength 2147483647
   */
  password: string;
  firstName: string;
  lastName: string;
  country: string;
  phone?: string;
}

export interface ApiResponseAuthResponseDTO {
  success?: boolean;
  code?: number;
  message?: string;
  data?: AuthResponseDTO;
}

export interface AuthResponseDTO {
  accessToken?: string;
  refreshToken?: string;
  accessTokenExpiresIn?: number;
  userInfo?: UserInfoDTO;
}

export interface UserInfoDTO {
  id?: number;
  email?: string;
  firstName?: string;
  lastName?: string;
}

export type ApiResponseVoidData = { [key: string]: unknown };

export interface ApiResponseVoid {
  success?: boolean;
  code?: number;
  message?: string;
  data?: ApiResponseVoidData;
}

export interface LoginRequestDTO {
  email: string;
  /**
   * @minLength 8
   * @maxLength 2147483647
   */
  password: string;
}

export interface ApiResponsePagedResponseDTOFlightResponseDTO {
  success?: boolean;
  code?: number;
  message?: string;
  data?: PagedResponseDTOFlightResponseDTO;
}

export interface FlightResponseDTO {
  id?: number;
  flightNumber?: string;
  departure?: string;
  destination?: string;
  departureDate?: string;
  departureTime?: string;
  price?: number;
}

export interface PagedResponseDTOFlightResponseDTO {
  content?: FlightResponseDTO[];
  pageNumber?: number;
  pageSize?: number;
  totalElements?: number;
  totalPages?: number;
  first?: boolean;
  last?: boolean;
}

export interface ApiResponseFlightResponseDTO {
  success?: boolean;
  code?: number;
  message?: string;
  data?: FlightResponseDTO;
}

export interface ApiResponsePagedResponseDTOBookingResponseDTO {
  success?: boolean;
  code?: number;
  message?: string;
  data?: PagedResponseDTOBookingResponseDTO;
}

export interface PagedResponseDTOBookingResponseDTO {
  content?: BookingResponseDTO[];
  pageNumber?: number;
  pageSize?: number;
  totalElements?: number;
  totalPages?: number;
  first?: boolean;
  last?: boolean;
}

export interface ApiResponseBoolean {
  success?: boolean;
  code?: number;
  message?: string;
  data?: boolean;
}

export interface AirportResponseDTO {
  id?: number;
  code?: string;
  name?: string;
  city?: string;
  country?: string;
}

export interface ApiResponseListAirportResponseDTO {
  success?: boolean;
  code?: number;
  message?: string;
  data?: AirportResponseDTO[];
}

export type GetBookingsParams = {
status?: string;
page?: number;
size?: number;
};

export type RefreshTokenBody = {[key: string]: string};

export type SearchFlightsParams = {
from: string;
to: string;
date: string;
page?: number;
size?: number;
};